tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_javascript_url_initial_insertion.html

<!DOCTYPE html>
<title>javascript: URL in iframe src, initial insertion check</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#process-the-iframe-attributes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
setup({single_test: true});
let iframeLoaded = false;
window.javascriptUrlRan = 0;
</script>
<iframe src="javascript:(() => { parent.javascriptUrlRan++; })()" onload="iframeLoaded = true; this.onload = assert_unreached;"></iframe>
<script>
onload = () => {
  const iframe = document.querySelector('iframe');
  assert_true(iframeLoaded, "iframeLoaded");
  iframe.src = iframe.src + ";";
  assert_equals(javascriptUrlRan, 1, "javascriptUrlRan");
  step_timeout(() => {
    assert_equals(javascriptUrlRan, 2, "javascriptUrlRan");
    done();
  }, 100); // Verify only one load event is fired on iframe
};
</script>
